Customizing eclipse build.
Putting elipse classes in eclipse-classes.  If they're
co-located with target, the two compilers (maven and eclipse)
step on each others' toes.  Also customizing the project names.

A bit of a mystery whether this needs to happen in all the
poms or just some of them, but this seems to work.