Lazily compute the StackFrame hashcode
Profiling of the stcap shows that ~45% of the memory is allocated in the
StackTraceWriterV2#write method.
The other 55% are mostly due to the JMX call, there's little we can do
here.

In the StackTraceWriterV2#write allocations, 80% come from the
StackFrame hashcode computation through the toString method: char buffer
+ String creation (which duplicate the array).

This patch lazily compute the StackFrame hashcode avoiding most of the
unneeded allocations.

In a simple stcap test, the memory allocated by StackTraceWriterV2#write
went from 1,27G to 130M